refactor(//py)!: Build system is now PEP517 compatible
BREAKING CHANGE: A ton of files have been moved in order to support
PEP517 standards based builds. Users are now advised to use commands
such as the following:

Editable installs:
python -m pip install --verbose --pre -e . [if necessary add --index-url
to access nightly builds of pytorch]

Wheel files:
python -m pip wheel --verbose --pre . [if necessary add --index-url
to access nightly builds of pytorch] --config-setting="--build-option=<options you would pass to setup.py>" -w dist

Alternative build processes are also supported

pypa/build:
python -m build --wheel --skip-dependency-check --no-isolation

> Skip dependency check means you need to install py/requirements.txt
first due to issues with TensorRT's python package


tools: adding a pyproject.toml linter


chore: update CI for new build path


Adding no deps flag

refactor: Formally define package contents in setup.py
