+++ py/ci/Dockerfile.ci
@@ -1,11 +1,11 @@
-FROM pytorch/manylinux-builder:cuda11.7
+FROM pytorch/manylinux-builder:cuda12.1
 ARG trt_version
 
 RUN echo -e "Installing with TensorRT ${trt_version}"
 
 RUN yum install -y ninja-build tensorrt-${trt_version}.*
 
-RUN wget https://github.com/bazelbuild/bazelisk/releases/download/v1.11.0/bazelisk-linux-amd64 \
+RUN wget https://github.com/bazelbuild/bazelisk/releases/download/v1.17.0/bazelisk-linux-amd64 \
     && mv bazelisk-linux-amd64 /usr/bin/bazel \
     && chmod +x /usr/bin/bazel
