+++ py/ci/build_manifest.txt
@@ -1,4 +1,5 @@
 py38
 py39
 py310
+py311
 libtorchtrt_pre_cxx11_abi
