+++ py/ci/soname_excludes.params
@@ -31,3 +31,9 @@
 --exclude libcublasLt.so.11
 --exclude libnvinfer.so.8
 --exclude libcudnn.so.8
+--exclude libcublas.so.12
+--exclude libcublasLt.so.12
+--exclude libcublas.so.12.1.3.1
+--exclude libcublasLt.so.12.1.3.1
+--exclude libcudart.so.11.8.89
+--exclude libcudart.so.11
