+++ py/torch_tensorrt/dynamo/conversion/impl/condition/ops.py
@@ -6,7 +6,7 @@ import torch
 from torch.fx.node import Target
 
 from torch_tensorrt.fx.types import TRTNetwork, TRTTensor
-from torch_tensorrt.dynamo.conversion import SourceIR
+from torch_tensorrt.dynamo._SourceIR import SourceIR
 from torch_tensorrt.dynamo.conversion.converter_utils import broadcastable
 from torch_tensorrt.fx.converters.converter_utils import (
     broadcast,
