+++ py/torch_tensorrt/dynamo/conversion/impl/embedding.py
@@ -8,10 +8,9 @@ import tensorrt as trt
 import torch
 from torch.fx.node import Target
 
+from torch_tensorrt.dynamo._SourceIR import SourceIR
 from torch_tensorrt.fx.types import TRTNetwork, TRTTensor
-
 from torch_tensorrt.fx.converters.converter_utils import (
-    SourceIR,
     set_layer_name,
 )
