+++ py/torch_tensorrt/dynamo/conversion/impl/normalization/ops.py
@@ -10,7 +10,7 @@ import logging
 
 from torch_tensorrt.fx.types import TRTNetwork, TRTTensor
 from torch_tensorrt.fx.utils import get_dynamic_dims
-from torch_tensorrt.dynamo.conversion import SourceIR
+from torch_tensorrt.dynamo._SourceIR import SourceIR
 
 from torch_tensorrt.fx.converters.converter_utils import (
     get_trt_plugin,
