+++ py/torch_tensorrt/dynamo/conversion/impl/permutation.py
@@ -4,7 +4,7 @@ from typing import Optional, Sequence, cast
 from torch.fx.node import Target
 
 from torch_tensorrt.fx.types import TRTNetwork, TRTTensor
-from torch_tensorrt.dynamo.conversion import SourceIR
+from torch_tensorrt.dynamo._SourceIR import SourceIR
 from torch_tensorrt.fx.converters.converter_utils import (
     set_layer_name,
     get_positive_dim,
