+++ py/torch_tensorrt/dynamo/conversion/impl/unary/base.py
@@ -7,7 +7,7 @@ from torch_tensorrt.fx.types import (
     TRTNetwork,
     TRTTensor,
 )
-from torch_tensorrt.dynamo.conversion import SourceIR
+from torch_tensorrt.dynamo._SourceIR import SourceIR
 from torch_tensorrt.fx.converters.converter_utils import set_layer_name
