--- py/versions.py
@@ -1,4 +1,0 @@
-__version__ = "2.0.0.dev0"
-__cuda_version__ = "12.1"
-__cudnn_version__ = "8.8"
-__tensorrt_version__ = "8.6"
