+++ toolchains/ci_workspaces/WORKSPACE.x86_64
@@ -36,7 +36,7 @@ http_archive(
 # External dependency for torch_tensorrt if you already have precompiled binaries.
 local_repository(
     name = "torch_tensorrt",
-    path = "/opt/circleci/.pyenv/versions/3.9.4/lib/python3.9/site-packages/torch_tensorrt"
+    path = "/opt/circleci/.pyenv/versions/3.10.9/lib/python3.10/site-packages/torch_tensorrt"
 )
 
 # CUDA should be installed on the system locally
@@ -65,13 +65,13 @@ new_local_repository(
 
 new_local_repository(
     name = "libtorch",
-    path = "/opt/circleci/.pyenv/versions/3.9.4/lib/python3.9/site-packages/torch",
+    path = "/opt/circleci/.pyenv/versions/3.10.9/lib/python3.10/site-packages/torch",
     build_file = "third_party/libtorch/BUILD"
 )
 
 new_local_repository(
     name = "libtorch_pre_cxx11_abi",
-    path = "/opt/circleci/.pyenv/versions/3.9.4/lib/python3.9/site-packages/torch",
+    path = "/opt/circleci/.pyenv/versions/3.10.9/lib/python3.10/site-packages/torch",
     build_file = "third_party/libtorch/BUILD"
 )
